moved _getPendingTransaction into the v0 implementation
